changed models

+++ Landlot.API/Models/LandGetResponseModel.cs
@@ -12,7 +12,7 @@ namespace Landlot.API.Models
         public Guid LandGuid { get; set; }
 
         [DataMember]
-        public int TotalArea { get; set; }
+        public decimal TotalArea { get; set; }
 
         [JsonConverter(typeof(LandlotMunicipalityConverter))]
         [DataMember]
@@ -48,7 +48,7 @@ namespace Landlot.API.Models
 
         public LandGetResponseModel() { }
 
-        public LandGetResponseModel(Guid guid, int area, LandlotMunicipality municipality, string estateNumber, LandlotCulture culture,
+        public LandGetResponseModel(Guid guid, decimal area, LandlotMunicipality municipality, string estateNumber, LandlotCulture culture,
                                     LandlotClass landclass, LandlotProcessing processing,LandlotProtectedZone zone,
                                     LandlotPropertyType property, LandlotDrainage drainage)
         {
