+++ Landlot.API/Models/LandPatchRequestModel.cs
@@ -10,7 +10,7 @@ namespace Landlot.API.Models
     public class LandPatchRequestModel
     {
         [DataMember]
-        public int? TotalArea { get; set; }
+        public decimal? TotalArea { get; set; }
 
         [JsonConverter(typeof(LandlotMunicipalityConverter))]
         [DataMember]
@@ -44,7 +44,7 @@ namespace Landlot.API.Models
         [DataMember]
         public LandlotDrainage? Drainage { get; set; }
 
-        public LandPatchRequestModel(int area, LandlotMunicipality municipality, string estateNumber, LandlotCulture culture,
+        public LandPatchRequestModel(decimal area, LandlotMunicipality municipality, string estateNumber, LandlotCulture culture,
                                     LandlotClass landclass, LandlotProcessing processing, LandlotProtectedZone zone,
                                     LandlotPropertyType property, LandlotDrainage drainage)
         {
